Move helper functions to util.dart
