+++ Assignment-1/lib/main.dart
@@ -1,17 +1,6 @@
 import 'dart:io';
 import 'item.dart';
-
-// helper function to throw exception when value is less than or equal to zero
-void isGreaterThanZero(num val) {
-  if (val <= 0) {
-    throw Exception("Value must be greater than zero.");
-  }
-}
-
-// helper function to check if given string is a number
-bool isNumeric(String s) {
-  return double.tryParse(s) != null;
-}
+import 'util.dart';
 
 void main() {
   // Continue taking inputs as long as user enters y/Y when asked.
